Step One: Introduce Outputs and "Fields" to Parameters
This PR is the first step in updating cnab-go to match the spec
with regard to Parameters and Outptus. The overall change is pretty large
when you consider adding JSON Schema, so I'm opening this PR first and will
do a follow-on with the JSON schema in the interest of making things easy to review.

Part of #17 and #14