Part 2: Introducing JSON Schema for Param and Outputs (#34)
* Part 2: Introducing JSON Schema for Param and Outputs

This is Part 2 of the change to introduce json schema support. This introduces a dependency on
https://github.com/qri-io/jsonschema and introduces a facade layer that hides that implementation detail
from the users of cnab-go. A struct was added, called Schema, to a new "definitions" package. This exposes
an exported type that covers all the JSON schema bits. This was desirable to shield end users of cnab-go from the implementation details of
a specific library and for some libraries, necessary because the generally used jsonschema libraries for Go either don't export the types
directly (https://github.com/xeipuuv/gojsonschema), export but in a harder to directly use way (https://github.com/qri-io/jsonschema). A third library
did check most of the boxes (https://github.com/santhosh-tekuri/jsonschema), but I felt like not exposing the choice directly to the end consumer
makes it easier to switch if needed.

The wrapping package currently uses (https://github.com/qri-io/jsonschema and marshalls into that schema type when a validation is requested. This also
removes old ParameterDefinition validation logic, as it is replaced by the use of json schema validation.

Part three will wire up all of this for outputs. They are currently present in structure, but there is no validation done in cnab-go for outputs (paths etc)

* Fix lint issues

* Added some comments

* Additional comments and validation error refactor

* Remove sensitive from outputs

* fixed lint error

* Code review items

* Additional review comments

* Updated error messages in GetType/GetTypes