Support tweaking the operation before running it
This will allow a tool to apply additional configuration, such as
injecting additional files or environment variables, before the
operation is run.

In our case, it will allow Porter to continue to implement dependencies
without having to make extensive changes to cnab-go because it can
inject the changes it requires through this configuration function
instead of modifying the cnab-go code directly.