Swap integer for float on the Schema struct (#116)
All bundles are ultimately represented in canonical json format, which
does not allow for decimal numbers, only integers. We are updating the
specification to remove numbers from the schema and only allow for
integers.

This updates our Schema struct to only use integers as well so that we
can only represent valid bundles which can be written with the canonical
json library that we use.