Add Docker driver config to keep container around
Currently, it is hard to debug the state of an invocation image after
the various files and environment variables have been inserted by the
runtime. By not cleaning up the container, we allow bundle developers
an easier way to debug problems.

The parameter is CLEANUP_CONTAINERS, default is true.

Fixes #128