Update crud.Store to support general storage
CRUD can be used to work with any type of item, not just claims. Update
the interface signature to support specifying what type of item is being
stored so that the controlling code, e.g. the claim store or eventually
a credential store, can say what type of data is being stored. For
example "claims" or "credentials".

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>