Pin golang.org/x/sys to fix windows compilation
Upstream golang.org/x/sys introduced a breaking change within a major
version, see https://github.com/golang/go/issues/34610. The
workaround is to pin it to the last revesion that is compatible with
the version of moby/moby that we are using.

Verify that the code compiles on windows, macos and linux during our CI
run to catch problems like this one (a regression in golang.org/x/sys)
in the future as soon as it happens.

Also do a reset of our go dependencies after bootstrap. When we use go
get to install global dependencies, it modifies our go.* files changing
the versions of the dependencies that we required, making the build
break, because kind requires different versions than docker. This resets
it back to what we need before we compile.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>