Add timestamps to CredentialSet
This allows tools to display when a credential set was created/modified
without relying upon file timestamps which aren't available when
credentials are stored in the generic crud datastore.

Since we have this data on Claim, it makes sense to have it on other
data that we generally store in the host environment too.

Signed-off-by: Carolyn Van Slyck <carolyn.vanslyck@microsoft.com>