Ensure a single connect per logical query (#224)
* Ensure a single connect per logical query

When the claimstore executes multiple queries against the crud store to
execute a logical query, e.g. ReadInstallationStatus, ensure that it
only results in a single call to Connect.

I've updated the claim store tests to use our MockStore so that we can
assert on connects/closes. The MockStore was using nested maps that was
hard to understand so I refactored it so it's easier to understand.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix setting MockStore count node name

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Alway sort in list operations

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add missing newline

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>