Fix error listing outputs when bundle definition changed
When an output was removed, an error was generated listing the outputs
for an installation:

could not determine if the output "user" is sensitive: output "user" not defined

After digging into the code where I traverse the last generated output
value, I found that I was capturing the value of a for loop variable
without first making it a local variable inside the for loop. Go doesn't
do closures the same way other languages do, and what this really ended
up doing was giving every instance the same reference. Not very useful.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>