Add schemaVersion to credentialSet
Add a schemaVersion field to credentialSet so that we can identify the
spec version that the credential set adheres to. This was very useful
for claims, we use it for data migrations when the claim spec changes,
and it is best to have it _before_ we need to change the spec. Worst
case we never change the spec and we all win.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>