Add schemaVersion to credentialSet (#229)
* Add schemaVersion to credentialSet

Add a schemaVersion field to credentialSet so that we can identify the
spec version that the credential set adheres to. This was very useful
for claims, we use it for data migrations when the claim spec changes,
and it is best to have it _before_ we need to change the spec. Worst
case we never change the spec and we all win.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix comment

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Add constructor for CredentialSet

* Ensure that SchemaVersion is set on new CredentialSets.
* Try a new pattern for the default schema that doesn't involve checking
for errors parsing the spec constant. We test the spec constant in our
tests and people can rely on us having set it properly.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>