Fix invocation image digest validation
Bundles pushed with cnab-to-oci will have invocation images with more
than one repoDigest, which is valid. One is for the original invocation
image that was pushed. The other is for the relocated invocation image
that is now inside the bundle repository.

https://porter.sh/distribute-bundles/#image-references-after-publishing

We should check all of the repoDigests and if a match is found, then it
is valid.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>