Write docker output to op writer
Instead of writing by default to os.Stdout, the docker driver writes to
op.Out and the new field op.Err. The op defaults those values to
stdout/stderr, so if you weren't setting them previously the behavior is
the same. But for tools like porter that were collecting the output,
this enables them to capture all the output and not have rogue lines
printed directly to the screen.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>