Handle reading undefined outputs
If an output was generated that isn't defined in the bundle, e.g. the
driver or tool added the output, then do not throw an error when it is
read. When reading through the claimstore only sensitive values are
encrypted, so if it doesn't have a definition, it was stored without
being encrypted.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>