Fix empty file extension bug in FileSystemStore
When listing documents in FileSystemStore with the file extension filter
set to "", any document with an extension was being filtered out. This
causes problems when an output (which doesn't require an extension) has
a period in the name, such as the logs output, or mydata.json.

I'v changed the logic so that when no filter is set for the file
extension, we return all files in the directory.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>