Do not remove file extension from output names
The FileSystemStore was removing the file extension from output names
but outputs can be generated with file extensions. So when there isn't a
file extension filter specified for an item, such as outputs, leave the
file extension as-is.

We will still remove the .json extension for documents such as claims.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>