Optionally save logs as output (#243)
* Optionally save logs as output

When action.SaveLogs is set, a copy of the invocation image logs are
saved as an output named io.cnab.outputs.invocationImageLogs.

The logs are copied as they are written to a logfile, which after the
bundle completes, are saved as an output when the claim result is saved.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Handle reading undefined outputs

If an output was generated that isn't defined in the bundle, e.g. the
driver or tool added the output, then do not throw an error when it is
read. When reading through the claimstore only sensitive values are
encrypted, so if it doesn't have a definition, it was stored without
being encrypted.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix empty file extension bug in FileSystemStore

When listing documents in FileSystemStore with the file extension filter
set to "", any document with an extension was being filtered out. This
causes problems when an output (which doesn't require an extension) has
a period in the name, such as the logs output, or mydata.json.

I'v changed the logic so that when no filter is set for the file
extension, we return all files in the directory.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Do not remove file extension from output names

The FileSystemStore was removing the file extension from output names
but outputs can be generated with file extensions. So when there isn't a
file extension filter specified for an item, such as outputs, leave the
file extension as-is.

We will still remove the .json extension for documents such as claims.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>

* Fix meatdata

I can't spell but we shall miss you meatdata...

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>