Change DefaultSchemaVersion to not return an error
There isn't really a need to force people to handle (or ignore) an error
from DefaultSchemaVersion. The default schema is hard coded and with tests
and buildtime asserts, we know that it was set to a valid value.

We were already throwing a panic or ignoring the error anyway at
runtime, so this is more of an assurance actually than what we had
before.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>