Add bundle.GetAction helper
Add GetAction function to bundle that helps retrieve the definition of
an action, including core actions: install, upgrade, uninstall. This
allows the client to use common logic for some checks about what
should happen during an action that is stateless or modifies bundle
resources.

For example, should a claim be recorded for a particular action or not.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>