Fix panic when op.Err is unset in Docker driver
The op.Err field was added after the docker driver was first written. I
noticed that when the caller of the docker driver doesn't set op.Err it
causes a panic when we attach to the container's logs.

This fixes our defaulting of where we write errors so that we are never
writing to a nil writer accidentally.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>