Support connecting to TLS secured docker host
When we create a docker cli instance using just the docker/cli library,
some important configuration doesn't happen, namely reading
DOCKER_TLS_VERIFY and DOCKER_CERT_PATH. Unlike DOCKER_HOST, these other
variables for connecting to a TLS secured docker host are not configured
in the main library functions but instead just in the CLI's main package
when the flags (for stuff like --tlsverify) are bound.

This means we need to configure this ourselves when using that library.
I've added a function that consolidates all the necessary configuration
steps to make a docker cli client that is configured the same as if
running the docker CLI directly on your computer.

I've tested this manually against a remote docker host secured with a
self-signed certificate and I'm able to build, push and run bundles on a
remote host with this fix.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>