Add github action that builds project
I've converted our Azure DevOps Pipeline to a GitHub Workflow.
I split out calling each target so that it's easier to find which step
failed and repeat just failing steps in the build.

This is using the newer setup-go github action that has built in support
for detecting which version of go to use based on the go.mod file and
also caching downloaded go mods to speed up subsequent builds.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>