Use the cnabio fork of image-relocation
The original repository at github.com/pivotal/image-relocation has been archived and
is not maintained. In order to keep that repo maintained, I have created a fork of it
in our github organization and am updating us to use the new location.

It is using a new go package name so that everyone down the dependency chain doesn't need
to maintain a replace statement forever pointing to the new location.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>